requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
blinker==1.4
Brlapi==0.8.2
certifi==2020.6.20
chardet==4.0.0
chrome-gnome-shell==0.0.0
click==7.1.2
colorama==0.4.4
command-not-found==0.3
cryptography==3.3.2
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distlib==0.3.3
distro==1.5.0
Django==3.1.13
evdev==1.4.0
filelock==3.3.0
hidpidaemon==18.4.6
httplib2==0.18.1
idna==2.10
jeepney==0.6.0
keyring==22.2.0
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.5
louis==3.16.0
macaroonbakery==1.3.1
mysql-connector-python==8.0.26
netifaces==0.10.9
numpy==1.21.2
oauthlib==3.1.0
pandas==1.3.3
pipenv==2021.5.29
platformdirs==2.4.0
pop-transition==1.1.2
protobuf==3.12.4
py4j==0.10.9
pyarrow==5.0.0
pycairo==1.16.2
pycups==2.0.1
pydbus==0.6.0
PyGObject==3.38.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.4.0
pyRFC3339==1.1
pyspark==3.1.2
python-apt===2.2.0-ubuntu0.21.04.1
python-dateutil==2.8.2
python-debian==0.1.39
python-xlib==0.29
pytz==2021.3
pyxdg==0.27
PyYAML==5.3.1
repolib==1.5.2
repoman==1.4.0
requests==2.25.1
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
sessioninstaller==0.0.0
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.2
system76driver==20.4.42
systemd-python==234
ubuntu-advantage-tools==27.2
ubuntu-drivers-common==0.0.0
ufw==0.36
urllib3==1.26.2
virtualenv==20.8.1
virtualenv-clone==0.5.7
wadllib==1.3.5
xdg==5
xkit==0.0.0
